+++ modulo4/projeto-labecommerce/request.rest
@@ -36,3 +36,7 @@ Content-Type: application/json
    "productId": "1662494145135",
    "quantity": 3
 }
+
+###
+
+GET http://localhost:3003/users/1662489473121/purchases
