+++ modulo4/projeto-labecommerce/src/app.ts
@@ -1,0 +1,16 @@
+import express, { Express, Request, Response } from 'express';
+import { AddressInfo } from 'net';
+import cors from 'cors';
+
+export const app: Express = express();
+app.use(express.json());
+app.use(cors());
+
+const server = app.listen(process.env.PORT || 3003, () => {
+    if (server) {
+        const address = server.address() as AddressInfo;
+        console.log(`Server is running in http://localhost:${address.port}`);
+    } else {
+        console.error(`Failure upon starting server.`);
+    }
+});;
