+++ modulo4/projeto-labecommerce/src/index.ts
@@ -1,28 +1,18 @@
-import express, { Express, Request, Response } from 'express';
-import { AddressInfo } from 'net';
-import cors from 'cors';
+import { app } from './app'
 import createUser from './endpoints/createUser';
 import getAllUsers from './endpoints/getAllUsers';
 import createProduct from './endpoints/createProduct';
 import getAllProducts from './endpoints/getAllProducts';
 import createPurchases from './endpoints/createPurchases';
+import getPurchaseByUserId from './endpoints/getPurchaseByUserId'
 
-const app: Express = express();
-app.use(express.json());
-app.use(cors());
 
 app.post('/user', createUser)
 app.post('/products', createProduct)
 app.post('/purchases', createPurchases)
 app.get('/users', getAllUsers)
 app.get('/products', getAllProducts)
+app.get("/users/:userId/purchases", getPurchaseByUserId)
 
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-        const address = server.address() as AddressInfo;
-        console.log(`Server is running in http://localhost:${address.port}`);
-    } else {
-        console.error(`Failure upon starting server.`);
-    }
-});; 
+
